Nurikabe, Light Up, and Tree Tent Bug Fixes (#542)
* Update NurikabeCell.java

Changed setType() to use element IDs rather than element names

* Renamed instances of Basic Rule to Direct Rule

* Renamed Basic Rule to Direct Rule for Tree Tent tests

* Renamed directories from Basic Rule to Direct Rules

* Renamed Direct Rules to Basic Rules in Light Up tests

* Added back BlockInVerticalPath
