+++ src/test/java/puzzles/treetent/rules/EmptyFieldDirectRuleTest.java
@@ -6,7 +6,7 @@ import edu.rpi.legup.puzzle.treetent.TreeTent;
 import edu.rpi.legup.puzzle.treetent.TreeTentBoard;
 import edu.rpi.legup.puzzle.treetent.TreeTentCell;
 import edu.rpi.legup.puzzle.treetent.TreeTentType;
-import edu.rpi.legup.puzzle.treetent.rules.LastCampingSpotDirectRule;
+import edu.rpi.legup.puzzle.treetent.rules.EmptyFieldDirectRule;
 import edu.rpi.legup.save.InvalidFileFormatException;
 import legup.MockGameBoardFacade;
 import legup.TestUtilities;
@@ -16,9 +16,9 @@ import org.junit.Test;
 
 import java.awt.*;
 
-public class LastCampingSpotBasicRuleTest {
+public class EmptyFieldDirectRuleTest {
 
-    private static final LastCampingSpotDirectRule RULE = new LastCampingSpotDirectRule();
+    private static final EmptyFieldDirectRule RULE = new EmptyFieldDirectRule();
     private static TreeTent treetent;
 
     @BeforeClass
@@ -29,15 +29,44 @@ public class LastCampingSpotBasicRuleTest {
 
     @Test
     public void EmptyFieldTest() throws InvalidFileFormatException {
-        TestUtilities.importTestBoard("puzzles/treetent/rules/LastCampingSpotBasicRule/LastCampingSpot", treetent);
+        TestUtilities.importTestBoard("puzzles/treetent/rules/EmptyFieldDirectRule/EmptyField", treetent);
         TreeNode rootNode = treetent.getTree().getRootNode();
         TreeTransition transition = rootNode.getChildren().get(0);
         transition.setRule(RULE);
 
         TreeTentBoard board = (TreeTentBoard) transition.getBoard();
 
-        TreeTentCell cell1 = board.getCell(1, 0);
-        cell1.setData(TreeTentType.TENT);
+        TreeTentCell cell1 = board.getCell(1, 1);
+        cell1.setData(TreeTentType.GRASS);
+
+        board.addModifiedData(cell1);
+
+        Assert.assertNull(RULE.checkRule(transition));
+
+        for (int i = 0; i < board.getHeight(); i++) {
+            for (int k = 0; k < board.getWidth(); k++) {
+                Point point = new Point(k, i);
+                if (point.equals(cell1.getLocation())) {
+                    Assert.assertNull(RULE.checkRuleAt(transition, board.getCell(k, i)));
+                }
+                else {
+                    Assert.assertNotNull(RULE.checkRuleAt(transition, board.getCell(k, i)));
+                }
+            }
+        }
+    }
+
+    @Test
+    public void DiagonalTreeTest() throws InvalidFileFormatException {
+        TestUtilities.importTestBoard("puzzles/treetent/rules/EmptyFieldDirectRule/DiagonalTree", treetent);
+        TreeNode rootNode = treetent.getTree().getRootNode();
+        TreeTransition transition = rootNode.getChildren().get(0);
+        transition.setRule(RULE);
+
+        TreeTentBoard board = (TreeTentBoard) transition.getBoard();
+
+        TreeTentCell cell1 = board.getCell(1, 1);
+        cell1.setData(TreeTentType.GRASS);
 
         board.addModifiedData(cell1);
 
@@ -56,6 +85,4 @@ public class LastCampingSpotBasicRuleTest {
         }
     }
 }
-
-
