+++ src/test/resources/puzzles/lightup/rules/BulbsInPathContradictionRule/BlockInVerticalPath
@@ -1,0 +1,11 @@
+<Legup>
+    <puzzle name="LightUp">
+        <board width="3" height="3">
+            <cells>
+                <cell value="-4" x="0" y="0"/>
+                <cell value="-1" x="0" y="1"/>
+                <cell value="-4" x="0" y="2"/>
+            </cells>
+        </board>
+    </puzzle>
+</Legup>
