+++ src/test/resources/puzzles/lightup/rules/EmptyCellinLightDirectRule/EmptyCells
