+++ src/test/resources/puzzles/lightup/rules/EmptyCornersDirectRule/EmptyCorners
