+++ src/test/resources/puzzles/lightup/rules/FinishWithBulbsDirectRule/FinishWithBulbs
