+++ src/test/resources/puzzles/lightup/rules/MustLightDirectRule/MustLight
