+++ src/test/resources/puzzles/treetent/rules/EmptyFieldDirectRule/EmptyField
