+++ src/test/resources/puzzles/treetent/rules/FinishWithGrassDirectRule/FinishWithGrass
