+++ src/test/resources/puzzles/treetent/rules/FinishWithTentsDirectRule/FinishWithTents
