+++ src/test/resources/puzzles/treetent/rules/LastCampingSpotDirectRule/LastCampingSpot
